requirements.txt
pip==22.3.1

numpy
pandas
networkx
pyyaml
poverty
seaborn
tqdm